Swap to tiktoken gpt2, train on tiny-shakespeare, bf16, bigger model
Tokenizer:
- tokenizer.py: replace from-scratch ByteLevelBPETokenizer with pretrained
  GPT-2 BPE via tiktoken (n_vocab=50257). Drop train()/load_or_train()
  and DEFAULT_PATH — nothing to train or save. encode() uses
  allowed_special="all" so a literal "<|endoftext|>" round-trips cleanly.
- pyproject.toml: tokenizers>=0.20 -> tiktoken>=0.7 (lockfile regenerated).
- main.py: Tokenizer.load_or_train(...) -> Tokenizer(). Prompt switched
  from Portuguese to "ROMEO:" to match the new corpus.

Data:
- input.txt now points at Karpathy's tiny-shakespeare (~1.1MB English).
  The stale pt-bpe.json artifact was removed.

Model (gpt.py):
- d_model 256 -> 384, block_size 128 -> 256, n_head stays at 8
  (head_dim=48). Total params 17.6M -> 30.0M; transformer body
  4.7M -> 10.6M, now matching nanoGPT shakespeare_char in body size.

Mixed precision:
- main.py: model.set_dtype(mx.bfloat16) right after lazy-param init.
  Halves weights/grads/optimizer-state memory and roughly doubles
  matmul throughput on Apple Silicon.
- train.py: loss_fn casts logits to fp32 before cross_entropy so the
  log-softmax stays numerically stable.

+++ train.py
@@ -42,7 +42,9 @@ def get_batch(data: mx.array, batch_size: int, block_size: int):
 # Training
 # ------------------------------------------------------------
 def loss_fn(model: GPT, x: mx.array, y: mx.array) -> mx.array:
-    logits = model(x)                                        # [B, T, V]
+    # Forward in the model's dtype (bf16 with mixed precision); cast logits to
+    # fp32 before cross-entropy for numerical stability of the log-softmax.
+    logits = model(x).astype(mx.float32)                     # [B, T, V]
     return nn.losses.cross_entropy(
         logits.reshape(-1, logits.shape[-1]),
         y.reshape(-1),
